Update API details
