+++ native/src/main/java/io/ballerina/lib/wso2/controlplane/ListenerArtifactHandler.java
@@ -31,17 +31,24 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
+import static io.ballerina.lib.wso2.controlplane.ArtifactUtils.LISTENER_NAMES_MAP;
 import static io.ballerina.lib.wso2.controlplane.ControlPlaneConstants.HOST;
+import static io.ballerina.lib.wso2.controlplane.ControlPlaneConstants.HTTP;
+import static io.ballerina.lib.wso2.controlplane.ControlPlaneConstants.HTTPS;
 import static io.ballerina.lib.wso2.controlplane.ControlPlaneConstants.HTTP_VERSION;
 import static io.ballerina.lib.wso2.controlplane.ControlPlaneConstants.INFERRED_CONFIG;
 import static io.ballerina.lib.wso2.controlplane.ControlPlaneConstants.LISTENERS;
 import static io.ballerina.lib.wso2.controlplane.ControlPlaneConstants.LISTENER_DETAIL;
+import static io.ballerina.lib.wso2.controlplane.ControlPlaneConstants.NAME;
 import static io.ballerina.lib.wso2.controlplane.ControlPlaneConstants.PACKAGE;
+import static io.ballerina.lib.wso2.controlplane.ControlPlaneConstants.PORT;
+import static io.ballerina.lib.wso2.controlplane.ControlPlaneConstants.PROTOCOL;
 import static io.ballerina.lib.wso2.controlplane.ControlPlaneConstants.REQUEST_LIMIT;
 import static io.ballerina.lib.wso2.controlplane.ControlPlaneConstants.REQUEST_LIMITS;
+import static io.ballerina.lib.wso2.controlplane.ControlPlaneConstants.SECURE_SOCKET;
 import static io.ballerina.lib.wso2.controlplane.ControlPlaneConstants.SERVICE;
 import static io.ballerina.lib.wso2.controlplane.ControlPlaneConstants.TIMEOUT;
-import static io.ballerina.lib.wso2.controlplane.Utils.getServiceListener;
+import static io.ballerina.lib.wso2.controlplane.Utils.getArtifact;
 
 /**
  * Native function implementations of the wso2 control plane module.
@@ -53,7 +60,8 @@ public class ListenerArtifactHandler {
         List<BListInitialValueEntry> artifactEntries = new ArrayList<>();
         Set<BObject> listeners = getNonDuplicatedListeners(ArtifactUtils.artifacts, currentModule);
         for (BObject listener : listeners) {
-            artifactEntries.add(ValueCreator.createListInitialValueEntry(getServiceListener(listener, currentModule)));
+            artifactEntries.add(ValueCreator.createListInitialValueEntry(
+                    getArtifact(LISTENER_NAMES_MAP.get(listener), currentModule)));
         }
         return artifactEntries;
     }
@@ -72,6 +80,9 @@ public class ListenerArtifactHandler {
     public BMap<BString, Object> getDetailedListener(BObject listener, Module currentModule) {
         Type originalType = listener.getOriginalType();
         BMap<BString, Object> listenerRecord = ValueCreator.createMapValue();
+        listenerRecord.put(StringUtils.fromString(NAME), StringUtils.fromString(LISTENER_NAMES_MAP.get(listener)));
+        listenerRecord.put(StringUtils.fromString(PROTOCOL), getListenerProtocol(listener));
+        listenerRecord.put(StringUtils.fromString(PORT), listener.get(StringUtils.fromString(PORT)));
         listenerRecord.put(StringUtils.fromString(PACKAGE),
                 StringUtils.fromString(originalType.getPackage().toString()));
         BMap<BString, Object> config = (BMap<BString, Object>)
@@ -89,6 +100,13 @@ public class ListenerArtifactHandler {
     private static BMap<BString, Object> getRequestLimit(BMap<BString, Object> config, Module module) {
         return ValueCreator.createReadonlyRecordValue(module, REQUEST_LIMIT,
                 (BMap<BString, Object>) config.getMapValue(StringUtils.fromString(REQUEST_LIMITS)));
+    }
+
+    private static BString getListenerProtocol(BObject listener) {
+        BMap<BString, Object> config = (BMap<BString, Object>)
+                listener.get(StringUtils.fromString(INFERRED_CONFIG));
+        Object secureSocket = config.get(StringUtils.fromString(SECURE_SOCKET));
+        return StringUtils.fromString(secureSocket == null ? HTTP : HTTPS);
     }
 
 }
